+++ OUDS/Core/Themes/OrangeCompact/Sources/OrangeCompactTheme.swift
@@ -148,7 +148,7 @@ public final class OrangeCompactTheme: OUDSTheme, @unchecked Sendable {
         let bulletList = OrangeCompactThemeBulletListComponentTokensProvider(spaces: spaces)
         let checkbox = OrangeCompactThemeCheckboxComponentTokensProvider(sizes: sizes, borders: borders)
         let chip = OrangeCompactThemeChipComponentTokensProvider(sizes: sizes, borders: borders, colors: colors, spaces: spaces, dimensions: dimensions)
-        let controlItem = OrangeCompactThemeControlItemComponentTokensProvider(sizes: sizes, borders: borders, colors: colors, spaces: spaces)
+        let controlItem = OrangeCompactThemeControlItemComponentTokensProvider(sizes: sizes, borders: borders, colors: colors, spaces: spaces, dimensions: dimensions)
         let divider = OrangeCompactThemeDividerComponentTokensProvider(borders: borders)
         let icon = OrangeCompactThemeIconComponentTokensProvider(colors: colors)
         let link = OrangeCompactThemeLinkComponentTokensProvider(sizes: sizes, colors: colors, spaces: spaces)