+++ CHANGELOG.md
@@ -26,6 +26,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `input tag` component with label medium moderate typography for text (Orange-OpenSource/ouds-ios#1450)
 - `control item` component token renamed `spacePaddingBlockDensityCompactTopAlignmentTopTextContainer`
 
+### Fixed
+
+- Duplicated dimensions tokens provider for themes definitions (Orange-OpenSource/ouds-ios#1539)
+
 ### Deprecated
 
 - `control item` component token `spacePaddingBlockDensityCompactTopAlignmentTopText_container`